+++ e_metrobus/navigation/stations.py
@@ -24,7 +24,7 @@ class Stations:
 
     @staticmethod
     def __calc_route_data(km, vehicle):
-        return vehicle.data * km / vehicle.passengers
+        return vehicle.data * km
 
     def get_stations(self):
         return [station for station in self.stations]
@@ -33,22 +33,22 @@ class Stations:
         return self.stations[from_station][to_station]
 
     def get_route_data(
-        self, from_station: str, to_station: str, vehicle=None
+        self, from_station: str, to_station: str
     ) -> Dict[str, DataPerKilometer]:
-        passenger_kilometer = self.get_distance(from_station, to_station)
+        distance = self.get_distance(from_station, to_station)
         return {
-            vehicle.name: self.__calc_route_data(passenger_kilometer, vehicle)
+            vehicle.name: self.__calc_route_data(distance, vehicle)
             for vehicle in VEHICLES
         }
 
     def get_route_data_for_vehicle(
         self, from_station: str, to_station: str, vehicle: str
     ) -> DataPerKilometer:
-        passenger_kilometer = self.get_distance(from_station, to_station)
+        distance = self.get_distance(from_station, to_station)
         i = 0
         while i < len(VEHICLES):
             if VEHICLES[i].name == vehicle:
-                return VEHICLES[i].data * VEHICLES[i].passengers * passenger_kilometer
+                return VEHICLES[i].data * distance
             i += 1
         raise KeyError("Vehicle not found", vehicle)